Handle enable and disable video differently.
Some USB cams are slow to spin up and some platforms don't handle
starting and stopping usb cams well. In those cases we want to keep
ffmpeg running and keep the camera operational. But we don't want to
send the video over the network and waste bandwidth when nobody is
viewing. This change allows us to enable and disable video as normal,
but keeps the USB/ffmpeg running.